Fix input color CI checks
